% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset}
\alias{Dataset}
\title{Get a target time series based on a covariate.}
\usage{
Dataset(start_date = "2020-01-01")
}
\arguments{
\item{start_date}{The date when your time series starts in text. Default: 2020-01-01.

Get campaign start date.}

\item{data_matrix}{A matrix which contains training data for causal impact. Default: NULL.}

\item{campaign_end}{The relative index in which the intervention ends. If it's 0, then the intervention will last until the end of total_duration. Default: 0.}

\item{as_date}{flag to specify if you want to get campaign_start as Date. Default: FALSE}

\item{campaign_start}{The days in integer on which the intervention has started as index from the beginning of given time series.}
}
\value{
The start date as text  (as_date=FALSE) or Date (as_date=TRUE).

Set start date

The campaign start date in integer (as_date=FALSE) or Date (as_date=TRUE).

Set campaign start date
}
\description{
Dataset class provides utility functions to deal with sample dataset.
}
